Remove `pip check`
Builds currently fail with this error:
```
+ pip check
pypardiso 0.4.0 requires mkl, which is not installed.
```
https://pypi.org/project/mkl/ and https://anaconda.org/conda-forge/mkl
don't have the same source and are not the same package. PyPardiso works
with both, however pip check will complain about not finding the pypi
package, even though the conda mkl is available.

In https://github.com/conda-forge/staged-recipes/pull/15106#discussion_r654696146,
it was suggested to use mkl-service as a workaround, but pypardsio drops
mkl-service as a dependency in v0.4.0, so this is no longer possible.